fix: iOS/Android demo build errors + ADR-010 compliance
- Fixed extractor.close() → extractor.release() across all files
- Fixed calibra.model imports → tona.model for migrated types
- Fixed API mismatches in agent-generated code:
  - TonalSegment: startTime→startSeconds, endTime→endSeconds, label nullable
  - AgilityScore: .score→.scores.first
  - VocalRangeResult: .low→.range.lower, .high→.range.upper
  - TesseraResult is non-optional
  - PitchHistogram/TonalSegment arrays non-optional
  - KotlinFloatArray→use Swift extensions (ADR-010 compliant)
  - ContourCleanup→PitchProcessingConfig
  - PitchContourExtractor.close()→release()
  - Int vs Int32 for KMP bridging
- Added import Combine to all new iOS ViewModels
- Added import VoxaTrace to IntonationView
- Added Package.local.swift for local dev workflow
- Both Android (compileDebugKotlin) and iOS (xcodebuild) build clean

+++ demo-apps/ios/VoxaTraceDemo/VoxaTraceDemo/Sections/VocalRange/VocalRangeViewModel.swift
@@ -127,7 +127,7 @@ final class VocalRangeViewModel: ObservableObject {
         audioTask = Task {
             let hwRate = AudioSessionManager.hardwareSampleRate
             for await buffer in recorder.audioBuffers {
-                newSession.addAudio(samples: buffer.samples, sampleRate: hwRate)
+                newSession.addAudio(samples: buffer.samples, sampleRate: Int32(hwRate))
             }
         }
     }
